Initial major commit of changes for NEsper 6.0.1.

+++ NEsper/NEsper/compat/threading/MonitorLock.cs
@@ -144,7 +144,7 @@ namespace com.espertech.esper.compat.threading
             return new TrackedDisposable(InternalRelease);
         }
 
-        public IDisposable Acquire(int msec)
+        public IDisposable Acquire(long msec)
         {
             InternalAcquire(msec);
             return new TrackedDisposable(InternalRelease);
@@ -158,7 +158,7 @@ namespace com.espertech.esper.compat.threading
         /// <param name="releaseLock"></param>
         /// <param name="msec"></param>
         /// <returns></returns>
-        public IDisposable Acquire(bool releaseLock, int? msec = null)
+        public IDisposable Acquire(bool releaseLock, long? msec = null)
         {
             InternalAcquire(msec ?? _uLockTimeout);
             if (releaseLock)
